rename json_string_binary -> json_stringn_nocheck
Summary:
This matches the name of the equivalent function that has since been
added in the upstream jansson library.

Add a helper function for producing a json_t from a w_string_t,
implemented in terms of json_stringn_nocheck, and make a sweep over the
code to find call sites that should use it.

Test Plan: `make integration`

Reviewers: sid0

Differential Revision: https://reviews.facebook.net/D29751