streams: add simple stream abstraction
Summary:
To reconcile the differences in IO between POSIX and Windows, this commit
introduces a simple abstraction for IO streams.  The abstraction defines some
basic read, write and seek primitives, and provides an event object that can be
signalled.  Both events and streams can be waited upon using appropriate kernel
primitives to avoid busy waiting.

The implementation for POSIX is pretty much a passthrough to the read(2),
write(2), lseek(2) and poll(2) system calls.  Events are implemented using the
pipe(2) system call.

For Windows, we use the system level ReadFile and WriteFile API.  Non-blocking
mode is implemented as overlapped IO, which queues IO to a system worker thread
and signals completion when the application enters an alertable wait state.
We use system Event handles for our event type, and use WaitForMultipleObjectsEx
to efficiently enter an alertable wait state.

This commit defines the interface, but does not adopt it; will do that in a
following diff.

Test Plan: `arc unit --everything` once whole stack lands :-/

Reviewers: sid0

Differential Revision: https://reviews.facebook.net/D32967