windows: add opendir/readdir
This implements opendir using the closest primitive available on
Windows, which works similarly to getdents().  The docs are unclear, but
it looks as though this requires Windows 8 or later.  The underlying
functionality is present at the NT layer in earlier versions should we
really need to support older versions of Windows.