windows: add widechar <-> utf-8 path mapping
Summary:
Rather than deal with potentially lossy ACP mapping, we're adopting
the stance that we are a unicode aware application when it comes
to interacting with the system, but we choose to represent our
strings as UTF-8 internally to minimize the portability overhead
of our Windows support.

To that end, where we call into filesystem related functions,
we will convert to the full extended UNC path name form (which unlocks
the maximum length filename possible) for the call.  When we enumerate
dirs or file change notifications, we convert from unicode to UTF-8.

Test Plan: arc unit

Reviewers: sid0

Differential Revision: https://reviews.facebook.net/D32985