windows: add broken lstat(2)
This is a simplistic implementation of lstat.  It doesn't follow reparse
points and attempts to populate the msvc struct stat in a mostly
compatible way to the MSVC stat() function.

Note that the MSVC runtime library has no way to formally recognize
symlinks, despite that support being present in Windows for several
years.  We report them as files because we cannot and don't want to
opendir them when we crawl the filesystem.

This isn't a great solution, but it is functional.