windows: add build and test plumbing
We're using good old nmake to perform our builds.
We use batch inference rules to invoke the compiler driver
once for each set (directory) of files that have changed.
This results in faster compilation times.

depends.js uses the /showIncludes flag in the compiler to extract
the include file dependency tree and generate dependency information,
similar to makedepend, so that we recompile the right things when
we change headers.

runtests.js is a very simple test driver that runs the unit tests
when `make check` is invoked.

We choose to use cscript with the JScript runtime for these because
it is less crazy-making to use than cmd.exe scripting and because
powershell cannot run scripts by default (which seems like a WTF
moment).

We still require arcanist to run the integration test suite.

We put the makefile in a subdir so that it is out of the way
from the unix build makefile, but we add a `make.bat` file in
the root of the repo so that we can simply run `make`, `make check`,
`make integration`, `make clean` and so on from the root.