windows: client: use named pipe to connect
Summary:
This makes use of the stream layer to connect to the watchman service.
We'll use a named pipe as this is the closest equivalent to a PF_LOCAL
available on Windows.  We construct a path specific to the user
and pass that through to our named pipe client stream implementation.

If we cannot connect, we'll spawn, using our implementation of posix_spawn,
the watchman service.

Test Plan: arc unit

Reviewers: sid0

Differential Revision: https://reviews.facebook.net/D33003