enable case insensitivity support (slightly) more generally
Summary:
Windows is also typically a case insensitive filesystem.  This expands
the name canonicalization checks that we had for OSX so that we can apply a
similar check/transformation on Windows.  This is also responsible for
rewriting short file names to long file names.

Musings: We should only observe short file names through injection from the
win32 dir watcher.  The name canonicalization executes a call to
FindFirstFile() under the covers, which is relatively expensive.  It would be
great to avoid this in the common case.  MSDN says that filesystems may have
mechanisms that encode short names in ways that are opaque to applications, so
we should not assume that 8.3 names containing tilde are the only way that a
short name may be represented.

We'll have to see how this performs in practice to see if we should consider
trying to optimize this.

Test Plan: arc unit

Reviewers: sid0

Differential Revision: https://reviews.facebook.net/D33015