osx: fixup some launchd warts
Summary:
On OSX, and perhaps some other environments, the TMPDIR
environment is randomized by default.  This means that in the default
install the watchman socket will be in a random location when you first
run it.  Since we capture this path in the launchd plist this can
result in not being able to communicate with watchman in cases where you
log in again and have a new randomized TMPDIR.

To avoid this issue we now default to always setting WATCHMAN_STATE_DIR
to a fixed and known default.  If not set explicitly at configure time,
we'll default to `$prefix/var/run/watchman` so that the default is
generally either `/usr/local/var/run/watchman` or `/var/run/watchman`
depending on whether this is a local or a system install.

The install rule for creating this dir has to make this dir world
writable so that any user can create their socket there.  This should be
no less secure than putting these files in `$TMPDIR`.

We also need to take some extra measures to ensure that launchd can pick
up these fixed paths; when we decide that we need to spawn watchman; we
repair it by first unloading an existing plist, then writing out the new
one with the newly configured settings.  This allows the repaired plist
to be loaded and the service to be started.  Without this, we would
rewrite the plist only, and launchd would punt on loading it since it
was already loaded.

https://github.com/facebook/watchman/issues/69

Test Plan:
`make integration` to prove that this didn't impact the integration
environment.  ad-hoc tests locally; `/usr/local/bin/watchman watch-list` to
start a pre-existing version of watchman without these changes, then
`./watchman watch-list` to run a version from this installation.

The local build is always able to replace and re-launch itself.  The old build
was not able to do this; it would get stuck for the timeout before reporting
that it was unable to talk to watchman.

Reviewers: durham, sid0

Reviewed By: sid0

Differential Revision: https://reviews.facebook.net/D32253