illumos: prefer portfs over inotify
Addresses https://github.com/facebook/watchman/issues/84

Newer Illumos based systems have an implementation of inotify that
doesn't match the Linux behavior, so if we are being built on a system
that has both portfs and inotify, we prefer to use portfs on the
assumption that it is the native implementation.

I doubt that we'll run into the opposite problem in the future; it seems
unlikely that Linux will add a portfs implementation and make it
available by default in the libc.