Clarify inotify related errors
This gets rid of the standard and misleading strerror translation of
errno values returned by inotify and replaces them with more directed
advice on how the user can resolve the problem.

We do this by allowing the watcher to pass in a more specific reason
than the standard strerror result, or NULL to use the strerror result.

https://github.com/facebook/watchman/issues/87