add `watch-project` command
Summary:
More discussion about this here:
https://github.com/ember-cli/ember-cli/issues/2643

The docs in the diff explain how this works and how to use it.

I'd like amasad and/or stefanpenner to review this from the
perspective of clients of the service.

Test Plan: `make integration`.  Docs tested using `cd website ; npm start` and following the newly added links.

Reviewers: amasad, sid0

Reviewed By: sid0

Subscribers: mostafaeweda

Differential Revision: https://reviews.facebook.net/D36117