watchman: increase fsevents latency parameter, make it tunable
Summary:
just taking a look at
https://github.com/ember-cli/ember-cli/issues/4101 and wondering what
tunings we actually have for fsevents, and it looks like just this
latency parameter.

The behavior of this parameter is:

> more than latency seconds have elapsed since the last event, your app
> will receive the event immediately. The delivery of the event resets
> the latency timer and any further events will be delivered after
> latency seconds have elapsed.

This diff makes the default latency larger, 0.01 seconds vs. the 0.0001
it was before, the idea being that we acheive a greater degree of
batching and reduce the change of `kFSEventStreamEventFlagUserDropped`
from happening.

Making the latency larger than this (say, 0.02 seconds) starts to have a
noticeable effect on the runtime of our integration tests, so 0.01 seems
to be a likely improved default.

We allow setting this latency parameter via both `.watchmanconfig` and
`/etc/watchman.json`.

Test Plan:
`make integration`.  I manually tried setting the latency in a watch
and observed the debug log line.

Reviewers: sid0

Reviewed By: sid0

Differential Revision: https://reviews.facebook.net/D38715