watchman: add defer_vcs boolean option for subscriptions
Summary:
The Nuclide folks need this so that they can be smarter around rebases
and other similar operations.

Note: I still need to update the docs for this, just wanted to diff this out
early.

Test Plan:
`make integration`.  Added a new test to show that we immediately
dispatch when `.hg/wlock` is changed.

Reviewers: sid0

Reviewed By: sid0

Differential Revision: https://reviews.facebook.net/D39009