Fixup edge case that was seemingly OS X specific
Nuclide wanted to subscribe to .hg/wlock file changes but would never
observe them in practice on OS X.

During debugging we noticed that if we touched the file and then
deleted it that the subscriptions would start to take effect.

It turns out that what was happening was that we'd get a notification
from the kernel, but by the time that we `lstat` the file it wasn't
present.  We had some logic to avoid creating a file node if the
file didn't exist.

In this case, because we'd never observed the file on the filesystem,
we'd never create a node for it.  This meant the its change was
effectively invisible as there was no node to move to the end of
the change list.

The fix is simple; we just create a deleted file node for this case.