Work around diverging and non-backwards compatible arc changes
Summary:
Recent API changes break our integration tests.  We can't simply
migrate to the new API because it is not compatible with the
version of arc that we use internally.

This diff is a quick stop-gap that allows us to continue to run our
tests without depending on the implementation of the arc/libphutil
test engine classes.

Longer term we'll remove this dependency completely.

Test Plan:
`make integration`.   You can also
`arc test tests/integration/absroot.php` to run individual test(s).

Reviewers: sid0

Reviewed By: sid0

Differential Revision: https://reviews.facebook.net/D39939