implement support for relative roots
Summary:
In general we want to discourage nested watches because they exhaust filesystem resources much more quickly.

However, in large multiproject repos, applications like 'buck' are interested in knowing what the state of the world is relative to a particular directory.

This diff implements support for relative roots to be specified during queries and subscriptions. These roots are expected to be subdirectories of the watched root. The contract for a relative root is straightforward:
- Any patterns and paths will be resolved relative to the relative root.
- Any paths returned will always be with respect to the relative root.
- No paths outside the relative root will be returned.

In other words, a relative root is exactly as if watchman were actually treating the specified directory as a root.

Triggers I'll do separately -- they need a little more work.

Test Plan: Test coverage.

Reviewers: wez

Reviewed By: wez

Differential Revision: https://reviews.facebook.net/D40467