watchman: add node bser module
Summary:
this is a javascript implementation of BSER encoding/decoding.

This helps us workaround the perf issues in
Refs: https://github.com/facebook/watchman/issues/113

I've performed only functional testing thus far, so I don't have any
real data on perf impact.  I'm hoping to avoid writing C++ to interface
with the Node/V8 internals.

For decoding, the strategy is to read the PDU length up-front,
accumulate the data into the buffer, and when we have the complete data,
recursively process and decode it.  This is much simpler than doing
something like a SAX streaming parser.  We can incrementally accumulate
the data so that we can plug into the node event loop.

The BunserBuf class will emit an event for each decoded value in the
stream.   A convenience method is provided that wraps around BunserBuf
and decodes an input string or Buffer and returns the result (rather
than emitting an event).

For encoding, to make things simpler, and because the input data tends
to be small and relatively simple, we only encode integers as int32
values.  This gives us a predictable size in the header region and
allows us to go back and fill in the size.

Since json numbers are all real values, all number values passed to the
BSER encoder will occupy 8 bytes.

Removed unused deps from the package, and bumped up the version number.

Test Plan:
`npm test`.

also:

```
cd watchman/node
npm link
cd ..
NODE_PATH=/usr/local/lib/node_modules node node/example.js
```

Reviewers: amasad

Reviewed By: amasad

Subscribers: sid0

Differential Revision: https://reviews.facebook.net/D40479