watch-project: re-use existing watch if one overlaps
Summary:
we missed this obvious-in-retrospect case when putting together
watch-project.

The problem scenario is this directory structure:

```
.watchmanconfig
subdir
   /.watchmanconfig
```

when `watch-project subdir` is invoked, it looks like a valid location
to start a watch, so we allow it.

This diff changes our logic so that we'll prefix match the existined
watched roots against the intended project path; if we find a match
we'll use that existing watch as the root.

If multiple roots are prefix matches, it is undefined which one we'll
take as the project root.

If no existing watch is found, we'll still choose to watch `subdir`
in this scenario.  Do we want to change the search logic to find the
"highest" instance of `.watchmanconfig` in the requested tree?

Test Plan: make integration

Reviewers: sid0

Differential Revision: https://reviews.facebook.net/D40653