windows: add Windows 7 port of GetOverlappedResultEx
This probes for GetOverlappedResultEx at runtime; if it is found we'll
use it.  Otherwise, we fall back to looping and using SleepEx.

It is important to trigger the alertable wait state otherwise the
WriteFileEx completion won't get triggered.

It is also important to not use set the wait parameter to
GetOverlappedResult because it will block the watchman service and
break concurrent clients.

Refs: https://github.com/facebook/watchman/issues/19