Log version and host name to log file, log all error responses
Summary:
some tools don't propagate error messages all to way to the
user and that makes it difficult to debug things just from the watchman
logs, so let's upgrade this from DEBUG level to ERROR level.

In addition, let's emit the version and build info to the logs when
we open them.

Test Plan: make integration.  Inspect logs to verify that both of these show up.

Reviewers: sid0

Reviewed By: sid0

Differential Revision: https://reviews.facebook.net/D43413