add capability checking to version command
Summary:
to make it easier for clients to optionally take advantage of new
features, this diff introduces the concept of capability names.

We add an optional second parameter to the `version` command that allows the
client to specify a list of optional and/or a list of required capability
names.  It is desirable to piggy back this on a version number check to allow
just one request/response round-trip to be issued to determine version
compatibility.  Thankfully, the version command from older servers doesn't
perform any rigorous checks on the number of parameters, so the extended
capabilities argument will be silently ignored.

The `version` command response will include a `capabilities` hash keyed by the
requested capability names with boolean values set to indicate whether the
server supports the named capability.

If a capability name is listed in the required set and the server does not
support it, an error response is returned.

The python and node clients are extended to add a `capabilityCheck` method.
The implementation adds some simple support for use against a server that does
not support capabilities; for a handful of capability names we know which
versions introduced them.  This allows clients to transition to using
capabilities before all servers have been upgraded.

Refs: https://github.com/facebook/watchman/issues/116

Test Plan: `make integration`

Reviewers: amasad, sid0

Reviewed By: sid0

Subscribers: stefanpenner

Differential Revision: https://reviews.facebook.net/D43935