website: fixup some js and css issues
Noticed that the side was blocked waiting for assets to load on my ipad.
Make a pass through and force the webfonts and highlighting stuff to
load async.

Ensure that our react stuff only tries to render after we've loaded
the react js library.

Break out of the nav for loop as soon as we've found a match (cuts build
time in half!)