node fb-watchman: update to 1.5.0
This performs more rigorous checks and error handling around spawning.
Previously we could fail to catch some classes of spawn failure.

We now try harder, and try to interpret and give useful advice in
a couple of cases.

We also suppress multiple 'error' events beyond the first during
spawning, as the first one is typically the most useful.

Also: we always log a spawn error, as some clients in common use
don't expose these errors and that makes it difficult for the
end user to understand what is going wrong.