more intelligent backoff during shutdown
At shutdown, we must wait for all clients to have been signalled
so that we can safely shutdown.  We we logging very frequently
while we waited, so this diff adds backoff and only logs when
the number of clients changes.

Later, we wait for all root objects to be freed.  It is possible
that a trigger child process will be still running for some time
and extend our lifetime as a result.  So let's give it an arbitrary
3 seconds before giving up.