introduce cmd_sync
Problem

When doing large changes to the filesystem with git, the amount of
changes received from the next watchman query take a long time to
transfer.  This is despite the fact that we know of all these changes
since we were the ones performing them.

Solution

Add a 'sync' command which syncs the root and returns a clock
representing the current time.  This is effectively fast-forwarding all
the updates we just performed.