Fix persistent client command if output pdu is different from server pdu
Summary:
Previously, when running:

  watchman --server-encoding=json --persistent log-level debug

many log messages sent by the server would be dropped.

I added lots of debugging, and found that whenever the server's PDU
was different from the client's PDU and we had to convert, we re-used
the same input buffer to convert the format to the client's format.

If there's only a single entry in the buffer, this was OK, but there
could be multiple entries in the buffer (or a partial entry at the
end of the buffer).

In the case of multiple entries in the buffer, re-using the same
input buffer to hold the results of the conversion meant we blew
away and lost the rest of the entries.

Since we could always have a partial entry at the end of the buffer,
I think we never want to re-use the input buffer for conversion.

This diff adds a new `output_pdu_buffer` which is used if necessary
to hold the results of the conversion to the client's format.

Test Plan:
`watchman --server-encoding=json --persistent log-level debug`

Previously, saw it dropped lots of log messages. Now, it shows them.

Reviewers: sid0, wez

Reviewed By: wez

Differential Revision: https://reviews.facebook.net/D44889