add separate io thread for stat/opendir operations
Summary:
the goal of this diff is to really dedicate the notify thread to just consuming
notifications and adding them to the pending queue, to minimize the risk of
triggering a recrawl.  We can't fully realize that with this diff as our big
root look is not granular enough, but it is a step towards this goal.

The actual filesystem operations are now processed in a new additional
IO thread.

As a side effect, we can now explicitly wait for the settle period to have
passed since we received the last kernel notification, which allows us to have
fewer wakeups when we're theoretically idle.

It also will make it easier to handle some aspects of symlink target change
notification handling that I've been brainstorming with @bhamilton, as well as some
possibly improved watcher fallback handling on osx if we detect brokenness in
fsevents (thinking of bringing back the background periodic crawl and refresh
that we used to to do, to deal with healing up after missing notifications).

Test Plan:
`make integration`.  Also inspect the process in activity monitor and
see that the idle wakeups are 0 compared to slowly increasing before this
build.

Reviewers: bhamilton, sid0

Reviewed By: sid0

Subscribers: bhamilton

Differential Revision: https://reviews.facebook.net/D44943