tidy up inotify move tracking
Summary: taking a new approach at this; now that we have decoupled the
pending notifications from the main structures in the root, there is a
greater chance for the root->dir->file data structures to drift away
from the notifications that come in more frequently and in smaller
batches.

The logical solution is for the watcher layer to maintain its mapping
of watch descriptors to names and pass on only name information to
the io thread and allow it to build up the tree state from there.

So this diff does that.   There is fan out for the other watchers
too; there are a couple of follow-on diffs to address that