Fixup a timing issue with triggers
Summary: travis was reliably failing the trigger test, but it was
relatively difficult to trigger locally.  The issue manifested when
we had a wakeup for the io thread from a "ping" that yielded no
pending items.

We were incorrectly jumping to immediately assessing triggers and
were not waiting for the settle period to pass.

This adds an explicit check for whether we were pinged or we timed
out when we sleep on the pending collection, so that we only enter
the trigger assessment path when the timeout was triggered.