Fixup `-j` option and multi-line json input
Summary: accidentally broke this when we refactored the guts for windows
support.  We previously allowed `-j` to read the whole of stdin and
parse it into a JSON value.  Since the windows code merge, we would
only read the first line.

This diff special cases json input when the stream is stdin; it sets
the mode to pretty json mode and we attempt to use the whole buffer
to process the input instead.