add helpful advice about tmux+launchd on OS X
A couple of folks have run into this, and the solution is not
especially clear; the `launchctl` command completes successfully
despite printing and error to stderr.

What we do here is, if we failed to connect to watchman on apple
systems, and TMUX is set in the environment, print out some advice
that can help remediate the situation.

Test Plan:  From outside tmux:

```
$ watchman shutdown-server
$ launchctl unload /Users/wez/Library/LaunchAgents/com.github.facebook.watchman.plist
$ tmux
 # now we're inside tmux
$ ./watchman version
/Users/wez/Library/LaunchAgents/com.github.facebook.watchman.plist: Operation not permitted
2015-09-17T17:41:58,224: [cli] unable to talk to your watchman on /opt/facebook/var/run/watchman/wez-state/sock! (Connection refused)
2015-09-17T17:41:58,225: [cli]
You may be hitting a tmux related session issue.
An immediate workaround is to run:

    watchman version

just once, from *outside* your tmux session, to allow the launchd
registration to be setup.  Once done, you can continue to access
watchman from inside your tmux sessions as usual.

Longer term, you may wish to install this tool:

    https://github.com/ChrisJohnsen/tmux-MacOSX-pasteboard

and configure tmux to use `reattach-to-user-namespace`
when it launches your shell.
```