use inotify_init1 to set CLOEXEC at init time
Summary: We have the potential to "leak" the inotify descriptor
to the child process if a watch is established at the same time as a
trigger is fired.  This slightly improves things if the kernel is new
enough.

We'd need something like a global mutex to completely eliminate that
possibility on all systems.