Avoid w_should_log_to_clients during W_LOG_FATAL
Summary: We've seen this get stuck spinning on OS X in a couple of crash
scenarios, so let's avoid calling it.  We don't yet know what the cause
of these crashes is, but we should at least be able to see it in the
logs once this diff is shipped.