Allow information to be passed from readdir to stat_path
If `pre_stat` is populated, it is a signal that we're walking a dir.
In this case we assume that the name returned from readdir is the
canonical case for the name of the file on disk and avoid making
an explicit call to resolve it.

There's a possible optimization here if we already know more about the
file being examined; if `pre_stat` is populated with stat information,
we can save an explicit lstat.  The consumer of the optimization is
present in the diff, but not the producer of the stat information.

I had to re-jigger how we call crawler to facilitate this; we need
to be able to directly call stat_path from crawler to pass the
`pre_stat` info.  This is means that we need to flip things around,
because we were previously calling crawler directly from stat_path.
Now we queue up calls to crawler from stat_path.  We do this by
extending the pending collection to signal that we want to invoke
just the crawler portion, so that we can avoid repeating or looping
over the same path forever.