Employ very strict O_NOFOLLOW semantics when we opendir
Summary: There are a couple of scenarios where a dir can be replaced
with a symlink that can skirt around our checks for symlinks.

We want to avoid traversing symlinks as this can be a costly operation
that takes us across filesystem boundaries.

This diff adds a stricter open operation that exclusively applies
O_NOFOLLOW to each component of the path that we want to open.
If any component of this fails, then the entire open operation fails.

We use this as part of our opendir_nofollow function.

We'll catch the failure to open in the various watcher
`root_start_watch_dir` functions, which then call `handle_open_errno`
and will recursively prune out the nodes from the tree.

Test Plan: `make integration` and run on my devserver.  Relatively lame.