pywatchman: add immutable version of data results
This adds an alternative object representation for decoded bser
values.   The new representation is immutable and can have its
properties accessed via either the mapping protocol or as attributes.

The goal is to eliminate an intermediate representation in mercurial
and just return the bser values directly in place of stat-like
objects that are manufactured by the hgwatchman extension.

This can save several seconds of runtime in hgwatchman if we adjust it
to process fresh_instance result sets with large trees.