pywatchman: add immutable version of data results
This adds an alternative object representation for decoded bser
values.   The new representation is immutable and can have its
properties accessed via either the mapping protocol or as attributes.

The goal is to eliminate an intermediate representation in mercurial
and just return the bser values directly in place of stat-like
objects that are manufactured by the hgwatchman extension.

This can save several seconds of runtime in hgwatchman if we adjust it
to process fresh_instance result sets with very large trees.

`bser.loads` has been augmented with an optional `mutable` flag that
defaults to `True`.   When `True` we return the same dicts and lists
of dicts that we've always done.  When `False` we return our immutable
object representation.

pybser has been updated to respect this flag also.

Test Plan: `make integration`.  I also built a debug cpython with ASAN
to verify the reference counting.