Add periodic socket liveness check
Summary: since we're using unix domain sockets, it is possible for a
couple of funky situations to happen:

* Someone can remove the socket file.  Perhaps they are being
  overzealous in fixing some issue, or perhaps the socket gets
  nuked by something like tmpwatch

* It's conceivable that there could be a race condition with multiple
  clients trying to start the service at the same time.

The latter is a bug that we should also try to run down and resolve,
but it is worthwhile to detect and auto remediate if we detect that
the unix socket is broken.

This diff will periodically (every 10 seconds, or each time `poll(2)`
errors out), schedule a thread that will connect and sanity check
that `watchman get-pid` returns our own pid.  If we have any failure
to connect or a mismatch in the result then we'll self terminate.

While working on this I realized that we broke W_LOG_FATAL at some
point; we were just skipping FATAL level messages because the value
was outside the interesting range(!).   That might explain some
of the weird CPU spins we've seen in a couple of weird cases, where
we should have terminated but are still trying hard.

This fixes the value so that it is below the "off" threshold.

Test Plan:

Run this in one window:

`./watchman --logfile=/dev/stdout --foreground -U /tmp/foow`

Then run the same thing in another window.

The first one will output:

```
./watchman --logfile=/dev/stdout --foreground -U /tmp/foow
2015-10-20T13:16:03,838: [listener] Watchman 4.0.0 93e07ce578178c781981960cd7ff0047626f4e19 starting up on wez-mbp
2015-10-20T13:16:03,838: [listener] Using watcher mechanism fsevents
2015-10-20T13:16:03,838: [listener] file limit is 2560 kern.maxfilesperproc=10240
2015-10-20T13:16:03,838: [listener] raised file limit to 10240
2015-10-20T13:16:03,838: [listener] launchd checkin failed: No such process
2015-10-20T13:16:43,853: [sockcheck] remote pid from get-pid (51812) doesn't match my pid (51718)
2015-10-20T13:16:43,854: [sockcheck] Fatal error detected at:
2015-10-20T13:16:43,854: [sockcheck] 0   watchman                            0x00000001045796d0 w_log + 672
2015-10-20T13:16:43,854: [sockcheck] 1   watchman                            0x00000001045790b0 check_my_sock + 432
2015-10-20T13:16:43,854: [sockcheck] 2   libsystem_pthread.dylib             0x00007fff909259b1 _pthread_body + 131
2015-10-20T13:16:43,854: [sockcheck] 3   libsystem_pthread.dylib             0x00007fff9092592e _pthread_body + 0
2015-10-20T13:16:43,854: [sockcheck] 4   libsystem_pthread.dylib             0x00007fff90923385 thread_start + 13
[1]    51718 abort      ./watchman --logfile=/dev/stdout --foreground -U /tmp/foow
```

I also tried removing `/tmp/foow` with a similar result.